+++ lib/src/string/checkers.dart
@@ -145,10 +145,6 @@ bool isSlug(String s) => Patterns.slug.hasMatch(s);
 /// Checks if the string is a valid hashtag (e.g., `#example`).
 bool isHashtag(String s) => Patterns.hashtag.hasMatch(s);
 
-/// Checks if the input string matches the given pattern.
-bool hasMatch(String pattern, String input, {bool caseSensitive = true}) =>
-    RegExp(pattern, caseSensitive: caseSensitive).hasMatch(input);
-
 /// Checks if the string is a valid latitude.
 bool isLatitude(String s) => Patterns.latitude.hasMatch(s);